requirements.txt
fastapi
uvicorn[standard]
pydantic
datasets
groq
python-dotenv

